+++ commands.go
@@ -148,7 +148,7 @@ func (c *AdminCommands) Say(args []string) (string, error) {
 	if err != nil {
 		return "", fmt.Errorf("usage: say [user] {message}: %w", err)
 	}
-	sendTTSFunc(user, msg)
+	sendTTSFunc(user, msg, c.cfg.TTSVoices())
 	sendNotificationFunc(user, msg)
 	return fmt.Sprintf("Sent to %s: %q", capitalize(user), msg), nil
 }